updated staff to unocss, still has axios error
Attempted to fix the table. It only gave a vague failed to add. After the attempt it gives this error. There may be an issue with the controller/dto set up:

JSON deserialization for type 'Selu383.SP25.P03.Api.Features.Staff.StaffDto' was missing required properties including: 'propertyName'., The dto field is required.